requirements.txt
Cython==0.29.24
pyjnius==1.3.0
matplotlib==3.4.3
numpyencoder==0.3.0
seaborn==0.11.2
pandas==1.3.2
scipy==1.7.1
torch==1.8.1
numpy==1.21.2
scikit_learn==0.24.2
tqdm==4.62.2
